Commit up to exercise 4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -27,6 +27,10 @@ blogsRouter.post( '/', async ( request, response ) => {
     return response.status( 400 ).end()
   }
 
+  if ( !request.token ) {
+    return response.status( 401 ).json( { error: 'token missing or invalid' } )
+  }
+
   const decodedToken = jwt.verify( request.token, process.env.SECRET )
   if ( !decodedToken.id ) {
     return response.status( 401 ).json( { error: 'token invalid' } )
@@ -47,6 +51,17 @@ blogsRouter.post( '/', async ( request, response ) => {
 })
 
 blogsRouter.delete( '/:id', async ( request, response ) => {
+  const blog = await Blog.findById( request.params.id )
+
+  if ( !request.token ) {
+    return response.status( 401 ).json( { error: 'token missing or invalid' } )
+  }
+
+  const decodedToken = jwt.verify( request.token, process.env.SECRET )
+  if ( !decodedToken.id || blog.user.toString() !== decodedToken.id ) {
+    return response.status( 401 ).json( { error: 'token invalid' } )
+  }
+
   await Blog.findByIdAndDelete( request.params.id )
   response.status( 204 ).end()
 })
